backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
uvicorn = {extras = ["standard"], version = "^0.30.0"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
sqlalchemy = "^2.0.42"
alembic = "^1.16.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.20"
pydantic-settings = "^2.10.1"
python-dotenv = "^1.1.1"
reportlab = "^4.0.0"
pandas = "^2.0.0"
qrcode = "^7.4.2"
pillow = "^10.0.0"
aiohttp = "^3.9.0"
websockets = "^12.0"
schedule = "^1.2.0"
openpyxl = "^3.1.0"
emails = "^0.6.0"
jinja2 = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"